[#44] Rename filter reducer to favorite
