+++ src/actions/favorite.js
@@ -1,0 +1,8 @@
+import * as types from './types'
+
+const toggleFavorite = id => ({
+  type: types.TOGGLE_FAVORITE,
+  id
+})
+
+export default toggleFavorite
