+++ src/containers/CompaniesScreen.js
@@ -4,9 +4,9 @@ import { loadCompanies, searchCompany } from '../actions/company'
 import CompaniesScreen from '../screens/Companies/CompaniesScreen'
 
 const mapStateToProps = state => ({
-  companyList: state.filterReducer.showFavorites
+  companyList: state.companyReducer.showFavorites
     ? state.companyReducer.items.filter(
-      company => state.filterReducer.favorites.indexOf(company.key) !== -1
+      company => state.favoriteReducer.favorites.indexOf(company.key) !== -1
     )
     : state.companyReducer.items,
   loading: state.companyReducer.loading,
