+++ src/reducers/index.js
@@ -1,10 +1,10 @@
 import { combineReducers } from 'redux'
 import companyReducer from './company'
 import eventReducer from './event'
-import filterReducer from './filter'
+import favoriteReducer from './favorite'
 
 export default combineReducers({
   companyReducer,
   eventReducer,
-  filterReducer
+  favoriteReducer
 })
